requirements.txt
discord.py
requests
python-dotenv
cryptography
openai
pytest
pytest-bdd
pydantic
pydantic_core
jitter
